+++ internal/application/agent_execution_policy.go
@@ -38,6 +38,7 @@ type AgentInvocationResolverV1 interface {
 type AgentTaskControlAuthorizationV1 struct {
 	TaskUUID string
 	Status   AgentTaskStatusV1
+	Workflow *AgentTaskWorkflowProjectionV1
 }
 
 type AgentTaskControlAuthorizerV1 interface {